prompt update, solutionWidget work, DataExtensions import
messing with the prompt, this one seems to be working ok currently which is better than the others as they had a high change on throwing exceptions. Prompt still needs thorough testing and refinement

did some with on the solutionDisplay widget, currently it's working as intended. It's not polished though, there will need to be some time for animations and such but the foundation is there

imported a dataextensions class to help with TryGet's from data dicts