HourglassController foundation, QOL updates
started the `HourglassController` class, currently it's only handling the timer text along with some core foundation for the timer in general. If time expires we highlight the remaining bubbles green.

Also fixed the bug where you could submit 2 of the same bubble

stubbed in some win/loss events and updating the solution target string on win/loss. This work needs to be refined but currently while working towards a complete game loop this will do.