finished number gen unit tests
finshed writing the test cases for the number gen, also added a little test with queueing results for future ideas.

played around with trying to run number gens on the thread pool, I wasn't able to sort it out but I also don't think this is an absolute necessity as running an individual number gen takes 12ms on my pc. This idea isn't lost though as we need to evaluate performance costs on low end devices so I'm leaving the code in as a breadcrumb