fixing RandomColorSelector
well I guess it's not so random anymore but that is not as important as the current implementation. Instead of selecting random colors from the array we now use a rotation approach with a queue. If we take the time to build out the colors array we can have a controlled solution that best displays this features goal by making sure we don't have colors too close together