fixed main thread bug with event
learned that we can switch back to the main thread in UniTasks, this fixes the bug where the listeners invoke main thread code. I don't see any drawback here as we swap to the main thread at the end of the task. Something new to keep in mind when this situation occurs again.

Also no longer raising an event from the constructor of `GameConfig`s, this was bad practice and should have never been done in the first place. A lot of UI only cares about number set data when the set has actually been created so hooking into that event serves the purpose and is also better quality code