added validation to GameConfig
added a validation step to the GameConfig constructor. Afaik I've covered all edge cases with the construction of a game config, I'm aware this is only supporting non duplicate allowed games of ints. Maybe to prepare for the future I should have an interface for GameConfigs and rename this config to "NonDuplicatesIntConfig"